Replaced type variable in LogRecord with AnyValue
